refactor: adapter 형태로 수정
