+++ src/main/java/in/koreatech/koin/domain/kakao/config/KakaoRequest.java
@@ -6,6 +6,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
+import in.koreatech.koin.domain.kakao.model.KakaoRequestType;
 import io.swagger.v3.oas.annotations.Hidden;
 
 @Hidden // Swagger 문서에 표시하지 않음
