+++ src/main/java/in/koreatech/koin/global/exception/KoinException.java
@@ -1,0 +1,10 @@
+package in.koreatech.koin.global.exception;
+
+public abstract class KoinException extends RuntimeException {
+
+    protected KoinException(String message) {
+        super(message);
+    }
+
+    public abstract String getDetail();
+}
